% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_les_donnees.R
\name{importe_les_donnees}
\alias{importe_les_donnees}
\title{importe_les_donnees Title}
\usage{
importe_les_donnees(date_debut, date_fin)
}
\arguments{
\item{date_debut}{une date en format yyyy-mm-dd}

\item{date_fin}{une date en format yyyy-mm-dd}
}
\value{
une chaine de caractère à coller dans le navigateur pour télécharger les données
}
\description{
importe_les_donnees Title
}
\examples{
importe_les_donnees(date_debut = "2022-01-01", date_fin = "2022-01-02")
}
